refactor: split workspace selectors to files
