+++ src/state/redux/workspaces/hooks.ts
@@ -4,6 +4,7 @@ import { Selector } from '@reduxjs/toolkit';
 import { useAppSelector } from '../hooks';
 import { selectWorkspace, WorkspaceData } from './workspaces';
 
+// Select profile and workspace from context
 export const useWorkspaceRootSelector = () => {
 	const workspace = 'default';
 	return useMemo(() => selectWorkspace(workspace), [workspace]);
